Removing masking node
In this version we remove the mask node to make the workflow cleaner. Future iterations need the ability to handle missing values in the test set (i.e., convert them to a class outside of the already given classes for pytorch to handle it). We also need to produce an image of the train set and validation set in epochs to check for over/under fitting.